Fixed duplicate question bug and autosave, secured form admin routes with server side auth
